+++ packages/evepilot-eve-ng/src/evepilot/eve_ng/__init__.py
@@ -1,0 +1,7 @@
+"""EVE-NG integration for EvePilot."""
+
+from evepilot.eve_ng.client import EveNgClient
+from evepilot.eve_ng.errors import EveNgError
+from evepilot.eve_ng.models import EveNgNode
+
+__all__ = ["EveNgClient", "EveNgError", "EveNgNode"]